passing floats as we now deal in mm

+++ energyPlotter/leonardoTests/datatype.h
@@ -3,7 +3,7 @@ typedef struct {
  // byte index;
   char id;
   char command;
-  int arg1;
-  int arg2;
+  float arg1;
+  float arg2;
 
 } Payload;
